[migration] extract test helper function into test-helper
